在打开GPS功能后，定位前，增加一条条GETGPS，再定位，就不会出现定位顺序错乱的问题。把定位失败后，状态调整到GPRS功能，不挂起线程，可以避免主线程标志位得不到清空，导致下次循环直接执行的问题。

+++ Hardware/Src/gprs.c
@@ -117,7 +117,7 @@ void GPRS_RstModule(void)
 /*******************************************************************************
  * gprs数据校验
  */
-uint8_t GPRS_VerifyCalculate(uint8_t* pBuffer, uint8_t size)
+uint8_t GPRS_VerifyCalculate(uint8_t* pBuffer, uint16_t size)
 {
 	uint8_t cal = 0;
